Backported vendor's rev2341, rev2428 and rev2639 into plugins' layouts.
The backported revisions are all affecting parts of the Redmine Core which are incorporated in the Collapse-plugin. These revisions are either merged into the 0.8-stable branch (though not requiring compatibility-checks) or aren't depending on changes in other parts of the Redmine Core which *are not* affected by this plugin.

This commit resolves issue #240.


git-svn-id: svn://evil-dev.net:3690/redmine-collapse/trunk@49 1809a232-8a5f-0410-be6b-a332444e9cb2