Fixed: tab-overflow when internationalized strings are longer than expected.
Slighty decreased the @padding@ of @#leftmenu-default #leftmenu .tabs ul li a@ CSS-selector from @padding: 4px 6px 4px 6px;@ to @padding: 4px 5px 4px 5px;@ to create enough space for the i18n'ed tab-names.

The problem came up with the German translation combined with the Default and the Alternate themes (not with Classic or Squeejee). Due to the lenghth of the German translations it results in the tabs overflowing it's dedicated "box".
This commit "fixes" this by slightly decreasing the tab-name's @padding@. Most-lekely this is something that has to be "fixed" again when more translations are added.

git-svn-id: svn://evil-dev.net:3690/redmine-collapse/trunk@156 1809a232-8a5f-0410-be6b-a332444e9cb2